+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
   * Adds storybook
   * Changes visuals of the service maturity sidebar to match OpsLevel
+  * Adds support for Scorecards
 
 ## [1.1.2] - 2023-10-24